requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
distribute==0.6.19
django-floppyforms==0.4.7
django-uni-form==0.8.0
wsgiref==0.1.2
